# Conflicts:
#	app/api/Datasets.scala
#	app/api/Files.scala
#	app/api/Indexes.scala
#	app/api/Metadata.scala
#	app/api/Status.scala
#	app/controllers/Files.scala
#	app/services/DI.scala
#	app/services/mongodb/MongoDBFileService.scala
#	app/services/mongodb/MongoDBMetadataService.scala
#	doc/src/sphinx/admin/customizing.rst
#	docker/play.plugins

+++ app/controllers/Admin.scala
@@ -20,7 +20,7 @@ import scala.concurrent.Future
  * Administration pages.
  */
 @Singleton
-class Admin @Inject() (sectionIndexInfo: SectionIndexInfoService, userService: UserService, metadataService: MetadataService) extends SecuredController {
+class Admin @Inject() (sectionIndexInfo: SectionIndexInfoService, userService: UserService, metadataService: MetadataService, versusService: VersusService) extends SecuredController {
 
   def customize = ServerAdminAction { implicit request =>
     val theme = AppConfiguration.getTheme
@@ -56,19 +56,13 @@ class Admin @Inject() (sectionIndexInfo: SectionIndexInfoService, userService: U
    * Gets the available Adapters from Versus
    */
   def getAdapters() = ServerAdminAction.async { implicit request =>
-    current.plugin[VersusPlugin] match {
-      case Some(plugin) => {
-        val adapterListResponse = plugin.getAdapters()
-        for {
-          adapterList <- adapterListResponse
-        } yield {
-          Ok(adapterList.json)
-        }
-      }
+    val versusEnabled : Boolean = play.api.Play.current.configuration.getBoolean("versusService").getOrElse(false)
 
-      case None => {
-        Future(Ok("No Versus Service"))
-      }
+    val adapterListResponse = versusService.getAdapters()
+    for {
+      adapterList <- adapterListResponse
+    } yield {
+      Ok(adapterList.json)
     }
   }
 
@@ -85,19 +79,16 @@ class Admin @Inject() (sectionIndexInfo: SectionIndexInfoService, userService: U
    * Gets available extractors from Versus
    */
   def getExtractors() = ServerAdminAction.async { implicit request =>
-    current.plugin[VersusPlugin] match {
-      case Some(plugin) => {
-        val extractorListResponse = plugin.getExtractors()
-        for {
-          extractorList <- extractorListResponse
-        } yield {
-          Ok(extractorList.json)
-        }
+    val versusEnabled : Boolean = play.api.Play.current.configuration.getBoolean("versusService").getOrElse(false)
+    if (versusEnabled) {
+      val extractorListResponse = versusService.getExtractors()
+      for {
+        extractorList <- extractorListResponse
+      } yield {
+        Ok(extractorList.json)
       }
-
-      case None => {
-        Future(Ok("No Versus Service"))
-      }
+    } else {
+      Future(Ok("VersusService not enabled"))
     }
   }
 
@@ -105,19 +96,16 @@ class Admin @Inject() (sectionIndexInfo: SectionIndexInfoService, userService: U
    * Gets available Measures from Versus
    */
   def getMeasures() = ServerAdminAction.async { implicit request =>
-    current.plugin[VersusPlugin] match {
-      case Some(plugin) => {
-        val measureListResponse= plugin.getMeasures()
-        for {
-          measureList <- measureListResponse
-        } yield {
-          Ok(measureList.json)
-        }
+    val versusEnabled : Boolean = play.api.Play.current.configuration.getBoolean("versusService").getOrElse(false)
+    if (versusEnabled) {
+      val measureListResponse= versusService.getMeasures()
+      for {
+        measureList <- measureListResponse
+      } yield {
+        Ok(measureList.json)
       }
-
-      case None => {
-        Future(Ok("No Versus Service"))
-      }
+    }  else {
+      Future(Ok("VersusService not enabled"))
     }
   }
 
@@ -125,19 +113,16 @@ class Admin @Inject() (sectionIndexInfo: SectionIndexInfoService, userService: U
    * Gets available Indexers from Versus
    */
   def getIndexers() = ServerAdminAction.async { implicit request =>
-    current.plugin[VersusPlugin] match {
-      case Some(plugin) => {
-        val indexerListResponse = plugin.getIndexers()
-        for {
-          indexerList <- indexerListResponse
-        } yield {
-          Ok(indexerList.json)
-        }
+    val versusEnabled : Boolean = play.api.Play.current.configuration.getBoolean("versusService").getOrElse(false)
+    if (versusEnabled) {
+      val indexerListResponse = versusService.getIndexers()
+      for {
+        indexerList <- indexerListResponse
+      } yield {
+        Ok(indexerList.json)
       }
-
-      case None => {
-        Future(Ok("No Versus Service"))
-      }
+    } else {
+      Future(Ok("VersusService not enabled"))
     }
   }
 
@@ -146,31 +131,29 @@ class Admin @Inject() (sectionIndexInfo: SectionIndexInfoService, userService: U
    * If an index has type and/or name, stores type/name in mongo db.
    */
    def createIndex() = ServerAdminAction.async(parse.json) { implicit request =>
-     current.plugin[VersusPlugin] match {
-       case Some(plugin) => {
-         Logger.trace("Contr.Admin.CreateIndex()")
-         val adapter =    (request.body \ "adapter").as[String]
-         val extractor =  (request.body \ "extractor").as[String]
-         val measure =    (request.body \ "measure").as[String]
-         val indexer =    (request.body \ "indexer").as[String]
-         val indexType =  (request.body \ "indexType").as[String]
-         val indexName =  (request.body \ "name").as[String]
-         //create index and get its id
-         val indexIdFuture: Future[models.UUID] = plugin.createIndex(adapter, extractor, measure, indexer)
-         //save index type (census sections, face sections, etc) to the mongo db
-         if (indexType != null && indexType.length !=0){
-          indexIdFuture.map(sectionIndexInfo.insertType(_, indexType))
-         }
-         //save index name to the mongo db
-         if (indexName != null && indexName.length !=0){
-          indexIdFuture.map(sectionIndexInfo.insertName(_, indexName))
-         }
-          Future(Ok("Index created successfully"))
-       }
 
-       case None => {
-         Future(Ok("No Versus Service"))
+     val versusEnabled : Boolean = play.api.Play.current.configuration.getBoolean("versusService").getOrElse(false)
+     if (versusEnabled) {
+       Logger.trace("Contr.Admin.CreateIndex()")
+       val adapter =    (request.body \ "adapter").as[String]
+       val extractor =  (request.body \ "extractor").as[String]
+       val measure =    (request.body \ "measure").as[String]
+       val indexer =    (request.body \ "indexer").as[String]
+       val indexType =  (request.body \ "indexType").as[String]
+       val indexName =  (request.body \ "name").as[String]
+       //create index and get its id
+       val indexIdFuture: Future[models.UUID] = versusService.createIndex(adapter, extractor, measure, indexer)
+       //save index type (census sections, face sections, etc) to the mongo db
+       if (indexType != null && indexType.length !=0){
+         indexIdFuture.map(sectionIndexInfo.insertType(_, indexType))
        }
+       //save index name to the mongo db
+       if (indexName != null && indexName.length !=0){
+         indexIdFuture.map(sectionIndexInfo.insertName(_, indexName))
+       }
+       Future(Ok("Index created successfully"))
+     } else {
+        Future(Ok("Index not created, VersusService not enabled"))
      }
    }
 
@@ -179,73 +162,68 @@ class Admin @Inject() (sectionIndexInfo: SectionIndexInfoService, userService: U
    * have type and/or name. Adds type and/or name to json object and calls view template to display.
    */
   def getIndexes() = ServerAdminAction.async { implicit request =>
-    current.plugin[VersusPlugin] match {
-      case Some(plugin) => {
-        Logger.trace(" Admin.getIndexes()")
-        val indexListResponse = plugin.getIndexes()
-        for {
-          indexList <- indexListResponse
-        } yield {
-          if(indexList.body.isEmpty())
-          {
-            Ok(Json.toJson(""))
-          }
-          else{
-            var finalJson: JsValue=null
-            val jsArray = indexList.json
-            //make sure we got correctly formatted list of values
-            jsArray.validate[List[VersusIndexTypeName]].fold(
-              // Handle the case for invalid incoming JSON.
-              // Note: JSON created in Versus IndexResource.listJson must have the same names as clowder models.VersusIndexTypeName
-              error => {
-                Logger.error("Admin.getIndexes - validation error")
-                InternalServerError("Received invalid JSON response from remote service.")
-                },
+    val versusEnabled : Boolean = play.api.Play.current.configuration.getBoolean("versusService").getOrElse(false)
+    if (versusEnabled) {
+      Logger.trace(" Admin.getIndexes()")
+      val indexListResponse = versusService.getIndexes()
+      for {
+        indexList <- indexListResponse
+      } yield {
+        if(indexList.body.isEmpty())
+        {
+          Ok(Json.toJson(""))
+        }
+        else{
+          var finalJson: JsValue=null
+          val jsArray = indexList.json
+          //make sure we got correctly formatted list of values
+          jsArray.validate[List[VersusIndexTypeName]].fold(
+            // Handle the case for invalid incoming JSON.
+            // Note: JSON created in Versus IndexResource.listJson must have the same names as clowder models.VersusIndexTypeName
+            error => {
+              Logger.error("Admin.getIndexes - validation error")
+              InternalServerError("Received invalid JSON response from remote service.")
+            },
 
-                // Handle a deserialized array of List[VersusIndexTypeName]
-                indexes => {
-                  val indexesWithNameType = indexes.map{
-                    index=>
-                        //check in mongo for name/type of each index
-                      val indType = sectionIndexInfo.getType(UUID(index.indexID)).getOrElse("")
-                      val indName = sectionIndexInfo.getName(UUID(index.indexID)).getOrElse("")
-                      //add type/name to index
-                      VersusIndexTypeName.addTypeAndName(index, indType, indName)
-                  }
-                  indexesWithNameType.map(i=> Logger.debug("Admin.getIndexes index with name = " + i))
-                  // Serialize as JSON, requires the implicit `format` defined earlier in VersusIndexTypeName
-                  finalJson = Json.toJson(indexesWithNameType)
-                }
-              ) //end of fold
-              Ok(finalJson)
-          }
+            // Handle a deserialized array of List[VersusIndexTypeName]
+            indexes => {
+              val indexesWithNameType = indexes.map{
+                index=>
+                  //check in mongo for name/type of each index
+                  val indType = sectionIndexInfo.getType(UUID(index.indexID)).getOrElse("")
+                  val indName = sectionIndexInfo.getName(UUID(index.indexID)).getOrElse("")
+                  //add type/name to index
+                  VersusIndexTypeName.addTypeAndName(index, indType, indName)
+              }
+              indexesWithNameType.map(i=> Logger.debug("Admin.getIndexes index with name = " + i))
+              // Serialize as JSON, requires the implicit `format` defined earlier in VersusIndexTypeName
+              finalJson = Json.toJson(indexesWithNameType)
+            }
+          ) //end of fold
+          Ok(finalJson)
         }
       }
-
-      case None => {
-        Future(Ok("No Versus Service"))
-      }
+    } else {
+      Future(Ok("No indices retrieved, VersusService not enabled"))
     }
+
   }
 
   /**
    * Builds a specific index in Versus
    */
   def buildIndex(id: String) = ServerAdminAction.async { implicit request =>
-    Logger.trace("Inside Admin.buildIndex(), index = " + id)
-    current.plugin[VersusPlugin] match {
-      case Some(plugin) => {
-      val buildResponse = plugin.buildIndex(UUID(id))
-        for {
-          buildRes <- buildResponse
-        } yield {
-          Ok(buildRes.body)
-        }
+    val versusEnabled : Boolean = play.api.Play.current.configuration.getBoolean("versusService").getOrElse(false)
+    if (versusEnabled) {
+      Logger.trace("Inside Admin.buildIndex(), index = " + id)
+      val buildResponse = versusService.buildIndex(UUID(id))
+      for {
+        buildRes <- buildResponse
+      } yield {
+        Ok(buildRes.body)
       }
-
-      case None => {
-        Future(Ok("No Versus Service"))
-      }
+    } else {
+      Future(Ok("VersusService not enabled, index not built."))
     }
   }
 
@@ -253,39 +231,34 @@ class Admin @Inject() (sectionIndexInfo: SectionIndexInfoService, userService: U
    * Deletes a specific index in Versus
    */
   def deleteIndex(id: String) = ServerAdminAction.async { implicit request =>
-    current.plugin[VersusPlugin] match {
-      case Some(plugin) => {
-        val deleteIndexResponse= plugin.deleteIndex(UUID(id))
-        for{
-          deleteIndexRes<-deleteIndexResponse
-        } yield {
-         Ok(deleteIndexRes.body)
-        }
+    val versusEnabled : Boolean = play.api.Play.current.configuration.getBoolean("versusService").getOrElse(false)
+    if (versusEnabled) {
+      val deleteIndexResponse= versusService.deleteIndex(UUID(id))
+      for{
+        deleteIndexRes<-deleteIndexResponse
+      } yield {
+        Ok(deleteIndexRes.body)
       }
+    } else {
+      Future(Ok("VersusService not enabled, index not deleted."))
+    }
 
-      case None => {
-        Future(Ok("No Versus Service"))
-      }
-		}
   }
 
   /**
    * Deletes all indexes in Versus
    */
   def deleteAllIndexes() = ServerAdminAction.async { implicit request =>
-    current.plugin[VersusPlugin] match {
-      case Some(plugin) => {
-        val deleteAllResponse = plugin.deleteAllIndexes()
-        for {
-          deleteAllRes <- deleteAllResponse
-        } yield {
-          Ok(deleteAllRes.body)
-        }
+    val versusEnabled : Boolean = play.api.Play.current.configuration.getBoolean("versusService").getOrElse(false)
+    if (versusEnabled) {
+      val deleteAllResponse = versusService.deleteAllIndexes()
+      for {
+        deleteAllRes <- deleteAllResponse
+      } yield {
+        Ok(deleteAllRes.body)
       }
-
-      case None => {
-        Future(Ok("No Versus Service"))
-      }
+    } else {
+      Future(Ok("VersusService not enabled, indices not deleted."))
     }
   }
